+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHB8DYZZKBXVPQ1ZQVKTTVHV"
+        "01KHBAQ506RGR9DMAFS3ENJRFP"
     )
     r"""The unique identifier of the data source"""