+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -134,7 +134,7 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-13T10:23:23.266Z")
+    created: Optional[datetime] = parse_datetime("2026-02-13T11:03:19.422Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")