+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHB8DYZWSWTYA5YQC2P4YFYV"
+        "01KHBAQ5028ACNNX0MM4MNMHV3"
     )
     r"""The unique identifier of the data source"""