+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-13T10:23:19.161Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-13T11:03:16.021Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")