+++ packages/orq-rc/RELEASES.md
@@ -5629,3 +5629,13 @@ Based on:
 - [python v4.3.0-rc.63] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.63] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.63 - packages/orq-rc
+
+## 2026-02-13 11:06:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.715.0 (2.822.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.64] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.64] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.64 - packages/orq-rc